pyproject.toml
[project]
name = "arl"
version = "0.1.0"
description = "AI Autonomous Research Lab - Multi-agent scientific research automation"
authors = [{name = "ARL Team"}]
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.0.0",
    "litellm>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "httpx>=0.25.0",
    "arxiv>=2.0.0",
    "PyPDF2>=3.0.0",
    "docker>=7.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.5.0",
    "black>=23.7.0",
    "ruff>=0.0.285",
]

[project.scripts]
arl = "arl.cli.main:cli"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["arl*"]
exclude = ["data*", "tests*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
